Undo powerline window focus "fix"
milkypostman/powerline#111 seems to work, but not always (e.g. on a
frame with multiple split windows, the powerlines don't get their focus
back right, apparently depending on mode.)  Anyway, I'm more comfortable
keeping the highlighted modeline where it is when window is out of
focus, as the cursor/point is a much more reliable indicator of that.